finished BIG work on normaliser, it works completely now... ...AND SHOULD BE USED ;-) TODO in here: a little bit debugging, maybe implement secGetline, secAbort!, code review (search for possible overflows), ...

+++ src/debug.h
@@ -1,8 +1,9 @@
 #ifndef __DEBUG_H__
 #define __DEBUG_H__
+#include <stdlib.h>
 #include "typedef.h"
 
-void debug(int type, const char *ptr, ...);
-void debugVerbose(int type, const char *ptr, ...);
+void debug(size_t type, const char *ptr, ...);
+void debugVerbose(size_t type, const char *ptr, ...);
 
 #endif
